+++ rio-backend/src/config/defaults.rs
@@ -64,7 +64,7 @@ pub fn default_working_dir() -> Option<String> {
 }
 
 #[inline]
-pub fn default_background_opacity() -> f32 {
+pub fn default_opacity() -> f32 {
     1.0
 }